+++ src/components/productList/CategoryItem.jsx
@@ -1,12 +1,8 @@
 import { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { Link, useNavigate } from 'react-router-dom';
-import { productActions } from '../../store/modules/productSlice';
+import { Link } from 'react-router-dom';
 
-const CategoryItem = ({ item, category }) => {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const { id, name, description, options, sales, inStock } = item;
+const CategoryItem = ({ item }) => {
+  const { id, name, description, options, inStock } = item;
   const price = options[0].price;
   const thumbImg = options[0].images.thumbnail.default;
   const hoverImg = options[0].images.thumbnail.hover;
