+++ src/components/productList/CategoryList.jsx
@@ -1,13 +1,10 @@
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import { productActions } from '../../store/modules/productSlice';
 import CategoryItem from './CategoryItem';
 
 const CategoryList = () => {
   const param = useParams();
-  const dispatch = useDispatch();
-  const { categoryData, loading } = useSelector((state) => state.product);
+  const { categoryData } = useSelector((state) => state.category);
   const { categoryName } = param;
 
   return (
